style: enhance ai gen
